2019 Day 9: Sensor Boost
Pretty simple day, but with a lot of changes everywhere. Adding a new parameter mode and instruction was trivial.

Making it possible to address memory outside the original program was easy as
well, just change List<Int> to Map<Int, Int> and modify the overloaded operators
for array access.

Changing all Ints to Longs was not difficult but required a lot of changes everywhere.

Next time there is a similar challenge I need to remember to start out with Longs and Maps from the beginning instead of Ints and Arrays.

+++ src/main/aoc2019/Day5.kt
@@ -3,16 +3,16 @@ package aoc2019
 
 class Day5(input: List<String>) {
 
-    private val parsedInput = input.map { it.toInt() }.toMutableList()
+    private val parsedInput = input.map { it.toLong() }.toMutableList()
 
 
-    fun solvePart1(): Int {
+    fun solvePart1(): Long {
         val computer = Intcode(parsedInput, mutableListOf(1))
         computer.run()
         return  computer.output.last()
     }
 
-    fun solvePart2(): Int {
+    fun solvePart2(): Long {
         val computer = Intcode(parsedInput, mutableListOf(5))
         computer.run()
         return  computer.output.last()
